Feat : CI 설정 (#122)
* Refactor: TestContainer로 Redis 띄움

* Feat: CI 추가

* Chore: 빌드 에러난 부분 Disabled 처리

* Chore: yml 레디스 설정 제거

* Chore: docker-compose test 관련 디비, 레디스 삭제

* Chore: 볼륨 빼먹음
