+++ build.gradle
@@ -78,11 +78,12 @@ dependencies {
     testImplementation "com.epages:restdocs-api-spec-mockmvc:${restdocsApiSpecVersion}"
     testImplementation "org.springframework.restdocs:spring-restdocs-mockmvc"
     testImplementation 'io.rest-assured:rest-assured'
+    testImplementation "org.testcontainers:testcontainers:1.19.0"
+    testImplementation "org.testcontainers:junit-jupiter:1.19.0"
 
     swaggerUI 'org.webjars:swagger-ui:4.11.1'
 
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.5'
-
 }
 
 test {
