+++ src/test/java/com/jelly/zzirit/ZziritApplicationTests.java
@@ -2,11 +2,11 @@ package com.jelly.zzirit;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
+
+import com.jelly.zzirit.global.redis.RedisTestContainerConfig;
 
 @SpringBootTest
-@ActiveProfiles("test")
-class ZziritApplicationTests {
+class ZziritApplicationTests extends RedisTestContainerConfig {
 
 	@Test
 	void contextLoads() {
