+++ src/test/java/com/jelly/zzirit/domain/cart/controller/CartControllerTest.java
@@ -18,7 +18,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 import org.springframework.restdocs.payload.FieldDescriptor;
-import org.springframework.test.context.ActiveProfiles;
 
 import com.jelly.zzirit.domain.cart.dto.response.CartItemResponse;
 import com.jelly.zzirit.domain.cart.dto.response.CartResponse;
@@ -32,7 +31,6 @@ import com.jelly.zzirit.global.support.RestDocsSupport;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 @Import(CartControllerTest.MockConfig.class)
-@ActiveProfiles("test")
 @Disabled
 class CartControllerTest extends RestDocsSupport {
