+++ src/test/java/com/jelly/zzirit/domain/member/controller/AuthControllerTest.java
@@ -3,12 +3,12 @@ package com.jelly.zzirit.domain.member.controller;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -16,11 +16,12 @@ import com.jelly.zzirit.domain.member.dto.request.EmailAuthDTO;
 import com.jelly.zzirit.domain.member.dto.request.EmailAuthVerificationDTO;
 import com.jelly.zzirit.domain.member.dto.request.SignupDTO;
 import com.jelly.zzirit.global.redis.RedisService;
+import com.jelly.zzirit.global.redis.RedisTestContainerConfig;
 
-@ActiveProfiles("test")
 @SpringBootTest
 @AutoConfigureMockMvc
-class AuthControllerTest {
+@Disabled
+class AuthControllerTest extends RedisTestContainerConfig {
 
 	@Autowired
 	MockMvc mockMvc;
