+++ src/test/java/com/jelly/zzirit/domain/member/service/email/EmailVerificationServiceTest.java
@@ -12,7 +12,6 @@ import org.springframework.test.context.ActiveProfiles;
 
 import com.jelly.zzirit.global.redis.RedisService;
 
-@ActiveProfiles("test")
 class EmailVerificationServiceTest {
 
 	@Mock
